pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "reposcale"
version = "0.1.0"
description = "Evaluation suite for codebase continuation and scaling intelligence"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["llm", "evaluation", "benchmark", "coding", "agents"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "jsonschema>=4.20",
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.1",
    "litellm>=1.40",
    "httpx>=0.27",
]

[project.scripts]
reposcale = "reposcale.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
